datausa_backend/requirements.txt
django
graphene-django
requests
